+++ air-script/tests/codegen/winterfell.rs
@@ -13,6 +13,26 @@ fn binary() {
 }
 
 #[test]
+fn bitwise() {
+    let generated_air = Test::new("tests/bitwise/bitwise.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../bitwise/bitwise.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
+fn buses_complex() {
+    let generated_air = Test::new("tests/buses/buses_complex.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../buses/buses_complex.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
 fn buses_simple() {
     let generated_air = Test::new("tests/buses/buses_simple.air".to_string())
         .transpile(Target::Winterfell)
@@ -21,6 +41,7 @@ fn buses_simple() {
     let expected = expect_file!["../buses/buses_simple.rs"];
     expected.assert_eq(&generated_air);
 }
+
 #[test]
 fn buses_simple_with_evaluators() {
     let generated_air = Test::new("tests/buses/buses_simple_with_evaluators.air".to_string())
@@ -32,12 +53,12 @@ fn buses_simple_with_evaluators() {
 }
 
 #[test]
-fn buses_complex() {
-    let generated_air = Test::new("tests/buses/buses_complex.air".to_string())
+fn buses_varlen_boundary_both() {
+    let generated_air = Test::new("tests/buses/buses_varlen_boundary_both.air".to_string())
         .transpile(Target::Winterfell)
         .unwrap();
 
-    let expected = expect_file!["../buses/buses_complex.rs"];
+    let expected = expect_file!["../buses/buses_varlen_boundary_both.rs"];
     expected.assert_eq(&generated_air);
 }
 
@@ -62,52 +83,33 @@ fn buses_varlen_boundary_last() {
 }
 
 #[test]
-fn buses_varlen_boundary_both() {
-    let generated_air = Test::new("tests/buses/buses_varlen_boundary_both.air".to_string())
-        .transpile(Target::Winterfell)
-        .unwrap();
+fn computed_indices_complex() {
+    let generated_air =
+        Test::new("tests/computed_indices/computed_indices_complex.air".to_string())
+            .transpile(Target::Winterfell)
+            .unwrap();
 
-    let expected = expect_file!["../buses/buses_varlen_boundary_both.rs"];
+    let expected = expect_file!["../computed_indices/computed_indices_complex.rs"];
     expected.assert_eq(&generated_air);
 }
 
 #[test]
-fn periodic_columns() {
-    let generated_air = Test::new("tests/periodic_columns/periodic_columns.air".to_string())
+fn computed_indices_simple() {
+    let generated_air = Test::new("tests/computed_indices/computed_indices_simple.air".to_string())
         .transpile(Target::Winterfell)
         .unwrap();
 
-    let expected = expect_file!["../periodic_columns/periodic_columns.rs"];
+    let expected = expect_file!["../computed_indices/computed_indices_simple.rs"];
     expected.assert_eq(&generated_air);
 }
 
 #[test]
-fn pub_inputs() {
-    let generated_air = Test::new("tests/pub_inputs/pub_inputs.air".to_string())
+fn constant_in_range() {
+    let generated_air = Test::new("tests/constant_in_range/constant_in_range.air".to_string())
         .transpile(Target::Winterfell)
         .unwrap();
 
-    let expected = expect_file!["../pub_inputs/pub_inputs.rs"];
-    expected.assert_eq(&generated_air);
-}
-
-#[test]
-fn system() {
-    let generated_air = Test::new("tests/system/system.air".to_string())
-        .transpile(Target::Winterfell)
-        .unwrap();
-
-    let expected = expect_file!["../system/system.rs"];
-    expected.assert_eq(&generated_air);
-}
-
-#[test]
-fn bitwise() {
-    let generated_air = Test::new("tests/bitwise/bitwise.air".to_string())
-        .transpile(Target::Winterfell)
-        .unwrap();
-
-    let expected = expect_file!["../bitwise/bitwise.rs"];
+    let expected = expect_file!["../constant_in_range/constant_in_range.rs"];
     expected.assert_eq(&generated_air);
 }
 
@@ -122,12 +124,21 @@ fn constants() {
 }
 
 #[test]
-fn constant_in_range() {
-    let generated_air = Test::new("tests/constant_in_range/constant_in_range.air".to_string())
-        .transpile(Target::Winterfell)
-        .unwrap();
+fn constraint_comprehension() {
+    let generated_air =
+        Test::new("tests/constraint_comprehension/constraint_comprehension.air".to_string())
+            .transpile(Target::Winterfell)
+            .unwrap();
 
-    let expected = expect_file!["../constant_in_range/constant_in_range.rs"];
+    let expected = expect_file!["../constraint_comprehension/constraint_comprehension.rs"];
+    expected.assert_eq(&generated_air);
+
+    let generated_air =
+        Test::new("tests/constraint_comprehension/cc_with_evaluators.air".to_string())
+            .transpile(Target::Winterfell)
+            .unwrap();
+
+    let expected = expect_file!["../constraint_comprehension/constraint_comprehension.rs"];
     expected.assert_eq(&generated_air);
 }
 
@@ -152,6 +163,16 @@ fn fibonacci() {
 }
 
 #[test]
+fn functions_complex() {
+    let generated_air = Test::new("tests/functions/functions_complex.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../functions/functions_complex.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
 fn functions_simple() {
     let generated_air = Test::new("tests/functions/functions_simple.air".to_string())
         .transpile(Target::Winterfell)
@@ -170,36 +191,6 @@ fn functions_simple_inlined() {
         .unwrap();
 
     let expected = expect_file!["../functions/functions_simple.rs"];
-    expected.assert_eq(&generated_air);
-}
-
-#[test]
-fn functions_complex() {
-    let generated_air = Test::new("tests/functions/functions_complex.air".to_string())
-        .transpile(Target::Winterfell)
-        .unwrap();
-
-    let expected = expect_file!["../functions/functions_complex.rs"];
-    expected.assert_eq(&generated_air);
-}
-
-#[test]
-fn variables() {
-    let generated_air = Test::new("tests/variables/variables.air".to_string())
-        .transpile(Target::Winterfell)
-        .unwrap();
-
-    let expected = expect_file!["../variables/variables.rs"];
-    expected.assert_eq(&generated_air);
-}
-
-#[test]
-fn trace_col_groups() {
-    let generated_air = Test::new("tests/trace_col_groups/trace_col_groups.air".to_string())
-        .transpile(Target::Winterfell)
-        .unwrap();
-
-    let expected = expect_file!["../trace_col_groups/trace_col_groups.rs"];
     expected.assert_eq(&generated_air);
 }
 
@@ -242,6 +233,26 @@ fn list_folding() {
         .unwrap();
 
     let expected = expect_file!["../list_folding/list_folding.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
+fn periodic_columns() {
+    let generated_air = Test::new("tests/periodic_columns/periodic_columns.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../periodic_columns/periodic_columns.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
+fn pub_inputs() {
+    let generated_air = Test::new("tests/pub_inputs/pub_inputs.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../pub_inputs/pub_inputs.rs"];
     expected.assert_eq(&generated_air);
 }
 
@@ -294,20 +305,31 @@ fn selectors_combine_with_list_comprehensions() {
 }
 
 #[test]
-fn constraint_comprehension() {
-    let generated_air =
-        Test::new("tests/constraint_comprehension/constraint_comprehension.air".to_string())
-            .transpile(Target::Winterfell)
-            .unwrap();
+fn system() {
+    let generated_air = Test::new("tests/system/system.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
 
-    let expected = expect_file!["../constraint_comprehension/constraint_comprehension.rs"];
+    let expected = expect_file!["../system/system.rs"];
     expected.assert_eq(&generated_air);
+}
 
-    let generated_air =
-        Test::new("tests/constraint_comprehension/cc_with_evaluators.air".to_string())
-            .transpile(Target::Winterfell)
-            .unwrap();
+#[test]
+fn trace_col_groups() {
+    let generated_air = Test::new("tests/trace_col_groups/trace_col_groups.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
 
-    let expected = expect_file!["../constraint_comprehension/constraint_comprehension.rs"];
+    let expected = expect_file!["../trace_col_groups/trace_col_groups.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
+fn variables() {
+    let generated_air = Test::new("tests/variables/variables.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../variables/variables.rs"];
     expected.assert_eq(&generated_air);
 }